pyproject.toml
[tool.poetry]
name = "climate-tracker"
version = "0.1.0"
description = "Climate Globe — interactive 3D globe with climate change data"
authors = []
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
sqlalchemy = "^2.0"
psycopg2-binary = "^2.9"
alembic = "^1.14"
pydantic-settings = "^2.7"
pgvector = "^0.3"
httpx = "^0.28"
pandas = "^2.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^6.0"
ruff = "^0.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
